+++ app/src/main/java/ru/stupnikov/application/data/Pouch.java
@@ -10,14 +10,14 @@ public class Pouch implements Serializable {
 
     public Pouch () {}
     public Pouch (String name) {this.name = name;}
-    public Pouch (String name, String valuta, int value, int position, ArrayList<String> listConvertableValuta) {
+    public Pouch (String name, String valuta, double value, int position, ArrayList<String> listConvertableValuta) {
         this.name =name; this.valuta = valuta; this.value =value; this.position =position;
         this.listConvertibleValuta = listConvertableValuta;
     }
 
     public String name = "default";
     public String valuta = "RUB";
-    public int value =0;
+    public double value = 0;
     public int position = 0;
     public ArrayList<String> listConvertibleValuta = new ArrayList<String>();
 }
